+++ packages/uni-h5/src/view/components/index.ts
@@ -1,6 +1,6 @@
 import Video from './video/index'
 import WebView from './web-view/index'
-import Map from './map/index'
+import Map from './map/index.vue'
 import CoverView from './cover-view'
 import CoverImage from './cover-image'
 import Picker from './picker/index'
